requirements.txt
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
cheroot==8.2.1
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
flask-whooshalchemy3==0.1.2
idna==2.8
jaraco.functools==2.0
Jinja2==2.10.3
#more-itertools==7.2.0
pbr==3.1.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.1.13
urllib3==1.25.6
Werkzeug==0.16.0
Whoosh==2.7.4
